Initial attempts to conjecturize Hypothesis
This is half in pieces and doesn't work very well yet. Some things
are working. Most aren't. It's really an initial proof of concept
more than it is even heading in the direction of a finished form.

It's also got some left over python 3 isms in it brought over from
Conjecture.

But:

1. Working shrink and generation engine.
2. Implementation of find on top of it
3. Implementation of some strategies.